fix (serde): item type can't be deserialized
To get good oapi schemagen a custom serializer was written. The
serializer outputs types that match the generated schema, but
the deserializer was generated and doesn't match this type.

This generates the serializer and deserializer then manually
implements the schema generation to get the correct reference types.

Testing showed that `skip_serializing_if="Option::is_none"` does not
work with `with="ts_seconds_option"`.m